+++ packages/abi/lib/interface.js
@@ -346,6 +346,7 @@ var Interface = /** @class */ (function () {
         }
         var bytes = (0, bytes_1.arrayify)(data);
         var reason = null;
+        var message = "";
         var errorArgs = null;
         var errorName = null;
         var errorSignature = null;
@@ -366,6 +367,12 @@ var Interface = /** @class */ (function () {
                     if (builtin.reason) {
                         reason = errorArgs[0];
                     }
+                    if (errorName === "Error") {
+                        message = "; VM Exception while processing transaction: reverted with reason string " + JSON.stringify(errorArgs[0]);
+                    }
+                    else if (errorName === "Panic") {
+                        message = "; VM Exception while processing transaction: reverted with panic code " + errorArgs[0];
+                    }
                 }
                 else {
                     try {
@@ -379,8 +386,9 @@ var Interface = /** @class */ (function () {
                 break;
             }
         }
-        return logger.throwError("call revert exception", logger_1.Logger.errors.CALL_EXCEPTION, {
+        return logger.throwError("call revert exception" + message, logger_1.Logger.errors.CALL_EXCEPTION, {
             method: functionFragment.format(),
+            data: (0, bytes_1.hexlify)(data),
             errorArgs: errorArgs,
             errorName: errorName,
             errorSignature: errorSignature,