+++ packages/ethers/lib/_version.d.ts
@@ -1,2 +1,2 @@
-export declare const version = "ethers/5.6.2";
+export declare const version = "ethers/5.6.3";
 //# sourceMappingURL=_version.d.ts.map