+++ packages/providers/lib.esm/json-rpc-provider.js
@@ -56,7 +56,7 @@ function checkError(method, error, params) {
         if (result) {
             return result.data;
         }
-        logger.throwError("missing revert data in call exception", Logger.errors.CALL_EXCEPTION, {
+        logger.throwError("missing revert data in call exception; Transaction reverted without a reason string", Logger.errors.CALL_EXCEPTION, {
             error, data: "0x"
         });
     }